refactor: move UpdateFiltersTrait to Helper and remove wasted call
- Move UpdateFiltersTrait from Site\View to Site\Helper namespace
  (shared utility code belongs with other helpers)
- Update imports in Cwmsermons and Cwmseriesdisplays HtmlViews
- Remove orphan Cwmimages::getShowHide() call in Cwmlandingpage
  HtmlView whose return value was discarded
